requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
loguru==0.7.2
python-dotenv==1.0.0
Pillow==10.4.0
groq==0.9.0
apscheduler==3.10.4
pytz==2024.1
pydantic==2.5.3
pydantic-core==2.14.6